Moved ConfigurationLoader to Controller as it needs to be used by the Model as well. Might be relevant to move it to the model and have it pass the configuration data down to the view in accordance with MVC.
Changed ConfigurationLoader into an utility class as it seems unnecessary to be creating instances of it, we are not changing its state during runtime.
